Moved salary and balance to account to be able to use it in admin aswell
