+++ src/main/account/Account.java
@@ -5,13 +5,17 @@ import main.Role;
 public abstract class Account {
     private int accountId;
     private static int idCounter = -1;
+    private int salary;
+    private int balance;
     private Enum<Role> role;
     private boolean login;
     private String username;
     private String password;
 
-    public Account(Enum<Role> role, String username, String password) {
+    public Account(int salary, int balance, Enum<Role> role, String username, String password) {
         this.accountId = idCounter++;
+        this.salary = salary;
+        this.balance = balance;
         this.role = role;
         this.login = false;
         this.username = username;
@@ -24,6 +28,22 @@ public abstract class Account {
 
     public void setAccountId(int accountId) {
         this.accountId = accountId;
+    }
+
+    public int getSalary() {
+        return salary;
+    }
+
+    public void setSalary(int salary) {
+        this.salary = salary;
+    }
+
+    public int getBalance() {
+        return balance;
+    }
+
+    public void setBalance(int balance) {
+        this.balance = balance;
     }
 
     public Enum<Role> getRole() {
