+++ src/main/admin/AdminModel.java
@@ -4,7 +4,7 @@ import main.Role;
 import main.account.Account;
 
 public class AdminModel extends Account {
-    public AdminModel(Enum<Role> role, String username, String password) {
-        super(role, username, password);
+    public AdminModel(int salary, int balance, Enum<Role> role, String username, String password) {
+        super(salary, balance, role, username, password);
     }
 }
