--- src/test/unit_account/CreateAccountTest.java
@@ -1,4 +1,0 @@
-package test.unit_account;
-
-public class CreateAccountTest {
-}
