+++ src/test/unit_admin/AdminControllerTest.java
@@ -1,0 +1,4 @@
+package test.unit_admin;
+
+public class AdminControllerTest {
+}
