--- src/test/unit_admin/AdminTest.java
@@ -1,4 +1,0 @@
-package test.unit_admin;
-
-public class AdminTest {
-}
