--- src/test/unit_admin/LoginAdminTest.java
@@ -1,8 +1,0 @@
-package test.unit_admin;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class LoginAdminTest {
-
-}
