+++ src/test/unit_user/UserControllerTest.java
@@ -2,12 +2,14 @@ package test.unit_user;
 
 import main.Role;
 import main.account.AccountList;
+import main.account.RemoveAccount;
 import main.user.UserController;
 import main.user.UserModel;
 import org.junit.Assert;
 import org.junit.Test;
 
 public class UserControllerTest {
+    RemoveAccount removeAccount = new RemoveAccount();
     UserModel user = new UserModel(1000, 0, Role.USER, "testuser", "testpassword");
     UserController userController = new UserController();
 
@@ -28,7 +30,7 @@ public class UserControllerTest {
     public void testRemoveAccount() {
         AccountList.accountArrayList.add(user);
         Assert.assertTrue("User not added to account list", AccountList.accountArrayList.contains(user));
-        userController.executeAccountDeletion(user,"testuser", "testpassword");
+        removeAccount.executeAccountDeletion(user,"testuser", "testpassword");
         Assert.assertFalse("User not removed from account list", AccountList.accountArrayList.contains(user));
     }
 }
